refactor(notifications): align prompt mobile wiring with DDD
