+++ CONTEXT.md
@@ -1,0 +1,18 @@
+# Context Glossary
+
+## NotificationPromptHistory
+
+Persisted history used to decide whether and when to show a notification opt-in prompt again.
+
+## NotificationPermissionStatus
+
+Domain-level notification permission status used by prompt eligibility rules, normalized from platform-specific permission APIs.
+
+## NotificationPromptPolicy
+
+Domain-level configuration that defines when notification prompts are eligible to appear.
+
+## NotificationPromptIntent
+
+User intent emitted from a notification prompt, such as accepting notifications or dismissing the prompt.
+
