+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/index.ts
@@ -1,37 +1,6 @@
 export { useNotifications } from "./hooks/useNotifications";
-export { NotificationsPromptWrapper } from "./new/NotificationsPromptWrapper";
-export { NotificationsPromptProvider } from "./new/NotificationsPromptProvider";
-export { useNotificationsContext } from "./new/NotificationsPromptProvider";
+export { NotificationsPromptWrapper } from "./components/NotificationsPromptWrapper";
+export { NotificationsPromptProvider } from "./components/NotificationsPromptProvider";
+export { useNotificationsContext } from "./components/NotificationsPromptProvider";
 export { useNotificationsData } from "./hooks/useNotificationsData";
-export { useNotificationsPrompt } from "./hooks/useNotificationsPrompt";
-export { useNotificationsDrawer } from "./hooks/useNotificationsDrawer";
-export {
-  INACTIVITY_DRAWER_DELAY_MS,
-  checkIsInactive,
-  evaluateAfterActionTrigger,
-  evaluateInactivityTrigger,
-  getNextRepromptDelay,
-  canPromptTransactionsAlertsForAction,
-  getNotificationPromptTarget,
-  shouldPromptOptInDrawerAfterAction,
-} from "./utils/notificationsPromptEngine";
-export type {
-  DataOfUser,
-  InitPushNotificationsDataResult,
-  NotificationCategory,
-  NotificationPromptTarget,
-} from "./types";
-export type {
-  AfterActionTriggerDecision,
-  InactivityTriggerDecision,
-  NotificationsPromptAfterActionSource,
-  NotificationsPromptRepromptDelay,
-  NotificationsPromptSkipDecision,
-  NotificationsPromptSkipReason,
-  NotificationsPromptShowDecision,
-  NotificationsPromptSource,
-} from "./utils/notificationsPromptEngine";
-export {
-  getPushNotificationsDataOfUserFromStorage,
-  setPushNotificationsDataOfUserInStorage,
-} from "./utils/storage";
+export type { InitPushNotificationsDataResult } from "./types";
