+++ domain/entity/notification-prompt/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./data/schema";
+export * from "./data/history";
 export * from "./data/slice";
 export * from "./logic/engine";
 export { notificationPromptReducer as default } from "./data/slice";
