+++ features/platform/notification-prompt/src/analytics/notificationPromptAnalytics.ts
@@ -1,17 +1,17 @@
-import { track } from "~/analytics";
-import {
-  type AfterActionTriggerDecision,
-  type InactivityTriggerDecision,
-  type NotificationPromptTarget,
+import type {
+  AfterActionTriggerDecision,
+  InactivityTriggerDecision,
+  NotificationPromptTarget,
 } from "@domain/entity-notification-prompt";
-import type { NotificationPromptTarget as MobileNotificationPromptTarget } from "../types";
 
 const GLOBAL_PUSH_NOTIFICATIONS_PROMPT_TARGET = "globalPushNotifications" as const;
 
-/** For drawer-visible events only; attempt events on skip keep drawerPromptTarget undefined. */
-export const resolveDrawerPromptTargetForAnalytics = (
-  drawerPromptTarget: MobileNotificationPromptTarget | undefined,
-): MobileNotificationPromptTarget => drawerPromptTarget ?? GLOBAL_PUSH_NOTIFICATIONS_PROMPT_TARGET;
+export const resolveDrawerPromptTargetForAnalytics = <
+  TPromptTarget extends string = NotificationPromptTarget,
+>(
+  drawerPromptTarget: TPromptTarget | undefined,
+): TPromptTarget | typeof GLOBAL_PUSH_NOTIFICATIONS_PROMPT_TARGET =>
+  drawerPromptTarget ?? GLOBAL_PUSH_NOTIFICATIONS_PROMPT_TARGET;
 
 const getDrawerPromptTargetFromAfterActionDecision = (
   decision: AfterActionTriggerDecision,
@@ -23,8 +23,9 @@ const getDrawerPromptTargetFromInactivityDecision = (
 ): NotificationPromptTarget | undefined =>
   decision.kind === "show" ? decision.drawerPromptTarget : undefined;
 
-export function trackAfterActionDecision(decision: AfterActionTriggerDecision) {
-  track("attempt_to_trigger_push_notification_drawer_after_action", {
+export const buildAfterActionDecisionAnalytics = (decision: AfterActionTriggerDecision) => ({
+  event: "attempt_to_trigger_push_notification_drawer_after_action" as const,
+  properties: {
     action: decision.source,
     shouldPrompt: decision.kind === "show",
     variant: decision.variant,
@@ -32,16 +33,17 @@ export function trackAfterActionDecision(decision: AfterActionTriggerDecision) {
     dismissedCount: decision.dismissedCount,
     skipReason: decision.kind === "skip" ? decision.reason : undefined,
     drawerPromptTarget: getDrawerPromptTargetFromAfterActionDecision(decision),
-  });
-}
+  },
+});
 
-export function trackInactivityDecision(decision: InactivityTriggerDecision) {
-  track("attempt_to_trigger_push_notification_drawer_after_inactivity", {
+export const buildInactivityDecisionAnalytics = (decision: InactivityTriggerDecision) => ({
+  event: "attempt_to_trigger_push_notification_drawer_after_inactivity" as const,
+  properties: {
     shouldPrompt: decision.kind === "show",
     variant: decision.variant,
     repromptDelay: decision.nextRepromptDelay,
     dismissedCount: decision.dismissedCount,
     skipReason: decision.kind === "skip" ? decision.reason : undefined,
     drawerPromptTarget: getDrawerPromptTargetFromInactivityDecision(decision),
-  });
-}
+  },
+});
