+++ features/platform/notification-prompt/src/index.ts
@@ -1,2 +1,4 @@
+export * from "./analytics/notificationPromptAnalytics";
 export * from "./permissions/mapFirebaseAuthorizationStatusToNotificationPermissionStatus";
 export * from "./policy/mapFeatureFlagsToNotificationPromptPolicy";
+export * from "./storage/notificationPromptStorage";
